requirements.txt
numpy
tqdm
Pillow>=8.0.0
matplotlib
paddlepaddle>=2.0.0
opencv-python>=4.5.0
visualdl
Flask>=2.0.0
